requirements.txt
google-generativeai
faiss-cpu
langchain
PyPDF2
python-dotenv
numpy
streamlit
tiktoken
